+++ src/main/java/de/ostfalia/se/converter/StaffConverter.java
@@ -15,7 +15,7 @@ public class StaffConverter implements Converter {
 
     @Override
     public Object getAsObject(FacesContext facesContext, UIComponent uiComponent, String s) {
-        return (ss.findById(Long.parseLong(s)));
+        return (ss.findById(Integer.parseInt(s)));
     }
 
     @Override
